+++ packages/ethereum-contracts/test/contracts/superfluid/SuperToken.NonStandard.test.ts
@@ -780,10 +780,9 @@ describe("SuperToken's Non Standard Functions", function () {
             );
 
             // specified spender is different than holder without allowance reverts
-            await expectRevertedWith(
-                customToken.callSelfTransferFrom(alice, bob, bob, 100),
-                "SuperToken: transfer amount exceeds allowance"
-            );
+            await expect(
+                customToken.callSelfTransferFrom(alice, bob, bob, 100)
+            ).to.be.revertedWithPanic("0x11");
 
             // holder must have enough balance
             await expectCustomError(