+++ packages/ethereum-contracts/test/foundry/superfluid/Superfluid.BatchCall.t.sol
@@ -45,33 +45,33 @@ contract SuperfluidBatchCallTest is FoundrySuperfluidTester {
 
     function test_Passing_If_Operation_Increase_Allowance_Is_Called_By_Host(
     ) public {
-        uint256 allowanceBefore = superToken.allowance(alice, bob);
+        uint256 aliceToBobAllowanceBefore = superToken.allowance(alice, bob);
 
         vm.prank(address(sf.host));
         superToken.operationIncreaseAllowance(alice, bob, 100);
-        uint256 allowanceAfter = superToken.allowance(alice, bob);
+        uint256 aliceToBobAllowanceAfter = superToken.allowance(alice, bob);
 
-        assertEq(allowanceAfter, allowanceBefore + 100);
+        assertEq(aliceToBobAllowanceAfter, aliceToBobAllowanceBefore + 100);
     }
 
     function test_Passing_If_Operation_Decrease_Allowance_Is_Called_By_Host(
     ) public {
-        uint256 allowanceBefore = superToken.allowance(alice, bob);
+        uint256 aliceToBobAllowanceBefore = superToken.allowance(alice, bob);
 
         vm.startPrank(address(sf.host));
         superToken.operationIncreaseAllowance(alice, bob, 100);
         superToken.operationDecreaseAllowance(alice, bob, 31);
         vm.stopPrank();
 
-        uint256 allowanceAfter = superToken.allowance(alice, bob);
-        assertEq(allowanceAfter, allowanceBefore + 69);
+        uint256 aliceToBobAllowanceAfter = superToken.allowance(alice, bob);
+        assertEq(aliceToBobAllowanceAfter, aliceToBobAllowanceBefore + 69);
     }
 
     function test_Passing_Increase_Allowance_Batch_Call(
         uint256 allowanceAmount
     ) public {
         ISuperfluid.Operation[] memory ops = new ISuperfluid.Operation[](1);
-        uint256 allowanceBefore = superToken.allowance(alice, bob);
+        uint256 aliceToBobAllowanceBefore = superToken.allowance(alice, bob);
         ops[0] = ISuperfluid.Operation({
             operationType: BatchOperation
                 .OPERATION_TYPE_ERC20_INCREASE_ALLOWANCE,
@@ -80,8 +80,8 @@ contract SuperfluidBatchCallTest is FoundrySuperfluidTester {
         });
         vm.prank(alice);
         sf.host.batchCall(ops);
-        uint256 allowanceAfter = superToken.allowance(alice, bob);
-        assertEq(allowanceAfter, allowanceBefore + allowanceAmount);
+        uint256 aliceToBobAllowanceAfter = superToken.allowance(alice, bob);
+        assertEq(aliceToBobAllowanceAfter, aliceToBobAllowanceBefore + allowanceAmount);
     }
 
     function test_Passing_Decrease_Allowance_Batch_Call(
@@ -93,7 +93,7 @@ contract SuperfluidBatchCallTest is FoundrySuperfluidTester {
         ISuperfluid.Operation[] memory ops = new ISuperfluid.Operation[](1);
         vm.prank(alice);
         superToken.increaseAllowance(bob, increaseAllowanceAmount);
-        uint256 allowanceBefore = superToken.allowance(alice, bob);
+        uint256 aliceToBobAllowanceBefore = superToken.allowance(alice, bob);
         ops[0] = ISuperfluid.Operation({
             operationType: BatchOperation
                 .OPERATION_TYPE_ERC20_DECREASE_ALLOWANCE,
@@ -102,14 +102,14 @@ contract SuperfluidBatchCallTest is FoundrySuperfluidTester {
         });
         vm.prank(alice);
         sf.host.batchCall(ops);
-        uint256 allowanceAfter = superToken.allowance(alice, bob);
-        assertEq(allowanceAfter, allowanceBefore - decreaseAllowanceAmount);
+        uint256 aliceToBobAllowanceAfter = superToken.allowance(alice, bob);
+        assertEq(aliceToBobAllowanceAfter, aliceToBobAllowanceBefore - decreaseAllowanceAmount);
     }
 
     function test_Passing_Increase_Decrease_Transfer_From_Batch_Call() public {
         ISuperfluid.Operation[] memory ops = new ISuperfluid.Operation[](2);
-        uint256 allowanceBefore = superToken.allowance(alice, bob);
-        uint256 balanceBefore = superToken.balanceOf(bob);
+        uint256 aliceToBobAllowanceBefore = superToken.allowance(alice, bob);
+        uint256 bobBalanceBefore = superToken.balanceOf(bob);
         ops[0] = ISuperfluid.Operation({
             operationType: BatchOperation
                 .OPERATION_TYPE_ERC20_INCREASE_ALLOWANCE,
@@ -125,20 +125,22 @@ contract SuperfluidBatchCallTest is FoundrySuperfluidTester {
 
         vm.prank(alice);
         sf.host.batchCall(ops);
+        uint256 aliceToBobAllowanceAfter = superToken.allowance(alice, bob);
+        assertEq(aliceToBobAllowanceAfter, aliceToBobAllowanceBefore + 50);
 
         vm.prank(bob);
         superToken.transferFrom(alice, bob, 50);
-        uint256 allowanceAfter = superToken.allowance(alice, bob);
         uint256 bobBalanceAfter = superToken.balanceOf(bob);
-        assertEq(allowanceAfter, 0);
-        assertEq(bobBalanceAfter, balanceBefore + 50);
+        aliceToBobAllowanceAfter = superToken.allowance(alice, bob);
+        assertEq(aliceToBobAllowanceAfter, 0);
+        assertEq(bobBalanceAfter, bobBalanceBefore + 50);
     }
 
     function test_Passing_Increase_Transfer_Allowance_And_Increase_Flow_Rate_Allowance()
         public
     {
         ISuperfluid.Operation[] memory ops = new ISuperfluid.Operation[](2);
-        uint256 allowanceBefore = superToken.allowance(alice, bob);
+        uint256 aliceToBobAllowanceBefore = superToken.allowance(alice, bob);
 
         (, , int96 flowRateAllowanceBefore) = sf.cfa.getFlowOperatorData(
             superToken,
@@ -152,7 +154,7 @@ contract SuperfluidBatchCallTest is FoundrySuperfluidTester {
             target: address(superToken),
             data: abi.encode(bob, 100)
         });
-        bytes memory callData = abi.encodeCall(
+        bytes memory increaseFlowRateAllowanceCallData = abi.encodeCall(
             IConstantFlowAgreementV1.increaseFlowRateAllowance,
             (superToken, bob, 100, new bytes(0))
         );
@@ -160,19 +162,19 @@ contract SuperfluidBatchCallTest is FoundrySuperfluidTester {
             operationType: BatchOperation
                 .OPERATION_TYPE_SUPERFLUID_CALL_AGREEMENT,
             target: address(sf.cfa),
-            data: abi.encode(callData, new bytes(0))
+            data: abi.encode(increaseFlowRateAllowanceCallData, new bytes(0))
         });
 
         vm.prank(alice);
         sf.host.batchCall(ops);
 
-        uint256 allowanceAfter = superToken.allowance(alice, bob);
+        uint256 aliceToBobAllowanceAfter = superToken.allowance(alice, bob);
         (, , int96 flowRateAllowanceAfter) = sf.cfa.getFlowOperatorData(
             superToken,
             alice,
             bob
         );
-        assertEq(allowanceAfter, 100);
+        assertEq(aliceToBobAllowanceAfter, aliceToBobAllowanceBefore + 100);
         assertEq(flowRateAllowanceAfter, flowRateAllowanceBefore + 100);
     }
 }