+++ packages/sdk-core/CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to the SDK-core will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+## Added
+- `increaseAllowance` and `decreaseAllowance` functions to SuperToken class
+- Batch call support for `increaseAllowance` and `decreaseAllowance`
+- `increaseFlowRateAllowance` and `decreaseFlowRateAllowance` functions to CFAv1 class
+
 ## [0.6.2] - 2023-02-07
 ### Added
 - `streamId` to `FlowUpdatedEvent` when querying from Subgraph